+++ backend/agentd/session.go
@@ -220,6 +220,11 @@ func (s *Session) Start() (err error) {
 	}()
 
 	for _, sub := range s.cfg.Subscriptions {
+		// Ignore empty subscriptions
+		if sub == "" {
+			continue
+		}
+
 		topic := messaging.SubscriptionTopic(org, env, sub)
 		logger.WithField("topic", topic).Debug("subscribing to topic")
 		subscription, err := s.bus.Subscribe(topic, agentID, s)